Updates to properly convert Synology Notes, including previously converted OneNote notebooks

+++ src/importer.py
@@ -88,6 +88,9 @@ class PathDeterminer:
             self.root_path = notebook.path
         for note in notebook.child_notes:
             note.path = notebook.path / common.safe_path(note.title)
+            if "1_1" in note.path.name:
+                note.path = notebook.path / common.safe_path(note.title.replace("1:1", "1-on-1"))
+                LOGGER.warning(f'Replaced name: {note.path.name}')
             # Don't overwrite existing suffices.
             if note.path.suffix != ".md":
                 note.path = note.path.with_suffix(note.path.suffix + ".md")
@@ -171,10 +174,15 @@ class FilesystemImporter:
         if "tags" in self.progress_bars:
             self.progress_bars["tags"].update(len(note.tags))
         assert note.path is not None
-        note.path.write_text(
-            note.get_finalized_body(self.include_title, self.frontmatter),
-            encoding="utf-8",
-        )
+        if note.path.exists():
+            #LOGGER.warning(f"File path exists, skipping: {note.path}")
+            return
+        else:
+            LOGGER.debug(f"Writing file: {note.path}")
+            note.path.write_text(
+                note.get_finalized_body(self.include_title, self.frontmatter),
+                encoding="utf-8",
+            )
 
     @common.catch_all_exceptions
     def import_note(self, note: imf.Note):
@@ -194,7 +202,9 @@ class FilesystemImporter:
         assert notebook.path is not None
         self.progress_bars["notebooks"].update(1)
         notebook.path.mkdir(exist_ok=True, parents=True)
-        for note in notebook.child_notes:
+        LOGGER.debug(f"Writing {len(notebook.child_notes)} notes from {notebook.title}")
+        for idx, note in enumerate(notebook.child_notes):
+            LOGGER.debug(f"Writing {idx}: {note.path}")
             self.import_note(note)
         for child_notebook in notebook.child_notebooks:
             self.import_notebook(child_notebook)
